Remove Credentials struct -  use the graphql Credentials type
