+++ README.md
@@ -203,18 +203,3 @@ Here are links to instructions for doing the CSV export:
 - [LastPass](https://support.lastpass.com/help/how-do-i-nbsp-export-stored-data-from-lastpass-using-a-generic-csv-file)
 - [1Password](https://support.1password.com/export/)
 - [Dashlane](https://support.dashlane.com/hc/en-us/articles/202625092-Export-your-passwords-from-Dashlane)
-
-## Roadmap
-
-### Next
-
-- [x] Add secure notes
-- [ ] Refactor: remove repetitive code in actions.rs and in graphql.rs
-- [ ] Refactor: Remove Credentials struct and only use the graphql Credentials type (similar to PaymentCards)
-
-### 3.0
-
-- [ ] Export of vault contents
-- [ ] push to vault from keychain
-- [ ] multiple users & vaults support ?
-- [ ] new vault items: payment cards, notes
