+++ TODO.md
@@ -1,0 +1,18 @@
+# TODO
+
+- test CSV import
+
+## Roadmap
+
+### Next
+
+- [x] Add secure notes
+- [ ] Refactor: remove repetitive code in actions.rs and in graphql.rs
+- [x] Refactor: Remove Credentials struct and only use the graphql Credentials type (similar to PaymentCards)
+
+### 3.0
+
+- [ ] Export of vault contents
+- [ ] push to vault from keychain
+- [ ] multiple users & vaults support ?
+- [ ] new vault items: payment cards, notes
