+++ src/graphql.rs
@@ -1,497 +1,13 @@
 use cynic::http::ReqwestExt;
 use reqwest::header;
 
+use crate::graphql::queries::types::*;
+
+pub mod queries;
+
 //const API_ENDPOINT: &str = "http://localhost:3000/api/graphql";
 const API_ENDPOINT: &str = "https://passlanevault.fly.dev/api/graphql";
 
-#[cynic::schema_for_derives(file = r#"src/schema.graphql"#, module = "schema")]
-pub mod queries {
-    use super::schema;
-    use crate::credentials::derive_encryption_key;
-    use crate::credentials::{decrypt, encrypt};
-    use core::fmt::Display;
-    use core::fmt::Formatter;
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct CredentialsQueryVariables {
-        pub grep: Option<String>,
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct EmptyQueryVariables {}
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(graphql_type = "Query", argument_struct = "CredentialsQueryVariables")]
-    pub struct MeQuery {
-        pub me: User,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(graphql_type = "Query", argument_struct = "EmptyQueryVariables")]
-    pub struct PaymentCardMeQuery {
-        pub me: UserWithPaymentCards,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(graphql_type = "Query", argument_struct = "EmptyQueryVariables")]
-    pub struct NotesMeQuery {
-        pub me: UserWithNotes,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(graphql_type = "Query", argument_struct = "EmptyQueryVariables")]
-    pub struct PlainMeQuery {
-        pub me: PlainUser,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "CredentialsQueryVariables")]
-    pub struct User {
-        pub auth_user_id: String,
-        pub created: Date,
-        pub email: String,
-        pub first_name: Option<String>,
-        pub last_name: Option<String>,
-        pub id: i32,
-        pub modified: Option<Date>,
-        pub vaults: Vec<Vault>,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "EmptyQueryVariables", graphql_type = "User")]
-    pub struct UserWithPaymentCards {
-        pub auth_user_id: String,
-        pub created: Date,
-        pub email: String,
-        pub first_name: Option<String>,
-        pub last_name: Option<String>,
-        pub id: i32,
-        pub modified: Option<Date>,
-        pub vaults: Vec<VaultWithPaymentCards>,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "EmptyQueryVariables", graphql_type = "User")]
-    pub struct UserWithNotes {
-        pub auth_user_id: String,
-        pub created: Date,
-        pub email: String,
-        pub first_name: Option<String>,
-        pub last_name: Option<String>,
-        pub id: i32,
-        pub modified: Option<Date>,
-        pub vaults: Vec<VaultWithNotes>,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "EmptyQueryVariables", graphql_type = "User")]
-    pub struct PlainUser {
-        pub auth_user_id: String,
-        pub created: Date,
-        pub email: String,
-        pub first_name: Option<String>,
-        pub last_name: Option<String>,
-        pub id: i32,
-        pub modified: Option<Date>,
-    }
-
-    impl PlainUser {
-        pub fn get_salt(&self) -> String {
-            format!("{}-{}", self.id, self.created).replace(":", "")
-        }
-
-        pub fn get_encryption_key(&self, master_pwd: &str) -> String {
-            let salt = self.get_salt();
-            derive_encryption_key(&salt, master_pwd)
-        }
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "CredentialsQueryVariables")]
-    pub struct Vault {
-        pub id: i32,
-        pub name: String,
-        #[arguments(grep = &args.grep)]
-        pub credentials: Option<Vec<Option<Credentials>>>,
-        pub personal: bool,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "EmptyQueryVariables", graphql_type = "Vault")]
-    pub struct VaultWithPaymentCards {
-        pub id: i32,
-        pub name: String,
-        pub payment_cards: Option<Vec<Option<PaymentCard>>>,
-        pub personal: bool,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(argument_struct = "EmptyQueryVariables", graphql_type = "Vault")]
-    pub struct VaultWithNotes {
-        pub id: i32,
-        pub name: String,
-        pub notes: Option<Vec<Option<Note>>>,
-        pub personal: bool,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    pub struct Credentials {
-        pub created: Date,
-        pub modified: Option<Date>,
-        pub id: i32,
-        pub password: String,
-        pub iv: Option<String>,
-        pub service: String,
-        pub username: String,
-    }
-
-    #[derive(cynic::Scalar, Debug, Clone)]
-    pub struct Date(pub String);
-
-    impl Display for Date {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{}", self.0)
-        }
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct AddGredentialsGroupMutationVariables {
-        pub input: AddCredentialsGroupIn,
-    }
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "AddGredentialsGroupMutationVariables"
-    )]
-    pub struct AddGredentialsGroupMutation {
-        #[arguments(input = AddCredentialsGroupIn {
-            credentials: args.input.credentials.clone(),
-            vault_id: args.input.vault_id
-        })]
-        pub add_credentials_group: i32,
-    }
-
-    #[derive(cynic::InputObject, Debug)]
-    pub struct AddCredentialsGroupIn {
-        pub credentials: Vec<CredentialsIn>,
-        pub vault_id: Option<i32>,
-    }
-
-    #[derive(cynic::InputObject, Debug, Clone)]
-    #[cynic(rename_all = "camelCase")]
-    pub struct CredentialsIn {
-        pub password_encrypted: String,
-        pub iv: String,
-        pub service: String,
-        pub username: String,
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct DeleteCredentialsMutationVariables {
-        pub input: DeleteCredentialsIn,
-    }
-
-    #[derive(cynic::InputObject, Debug, Clone)]
-    #[cynic(rename_all = "camelCase")]
-    pub struct DeleteCredentialsIn {
-        pub grep: String,
-        pub index: Option<i32>,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "DeleteCredentialsMutationVariables"
-    )]
-    pub struct DeleteCredentialsMutation {
-        #[arguments(input = DeleteCredentialsIn {
-            grep: args.input.grep.clone(),
-            index: args.input.index
-        })]
-        pub delete_credentials: i32,
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct DeletePaymentCardMutationVariables {
-        pub id: i32,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "DeletePaymentCardMutationVariables"
-    )]
-    pub struct DeletePaymentCardMutation {
-        #[arguments(id = args.id)]
-        pub delete_payment_card: i32,
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct MigrateMutationVariables {
-        pub new_key: String,
-        pub old_key: String,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "MigrateMutationVariables"
-    )]
-    pub struct MigrateMutation {
-        #[arguments(new_key = args.new_key.clone(), old_key = args.old_key.clone())]
-        pub migrate: i32,
-    }
-    #[derive(cynic::InputObject, Debug, Clone)]
-    pub struct ExpiryIn {
-        pub month: i32,
-        pub year: i32,
-    }
-    #[derive(cynic::QueryFragment, Debug, Clone)]
-    pub struct Expiry {
-        pub month: i32,
-        pub year: i32,
-    }
-    impl Display for Expiry {
-        fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-            write!(f, "{}/{}", self.month, self.year)
-        }
-    }
-
-    #[derive(cynic::InputObject, Debug, Clone)]
-    pub struct AddressIn {
-        pub street: String,
-        pub city: String,
-        pub country: String,
-        pub state: Option<String>,
-        pub zip: String,
-    }
-    impl AddressIn {
-        pub fn encrypt(&self, key: &str, iv: &str) -> AddressIn {
-            AddressIn {
-                street: encrypt(key, iv, &self.street),
-                city: encrypt(key, iv, &self.city),
-                country: encrypt(key, iv, &self.country),
-                state: if let Some(state) = &self.state {
-                    Some(encrypt(key, iv, state))
-                } else {
-                    None
-                },
-                zip: encrypt(key, iv, &self.zip),
-            }
-        }
-    }
-
-    #[derive(cynic::QueryFragment, Debug, Clone)]
-    pub struct Address {
-        pub id: i32,
-        pub street: String,
-        pub city: String,
-        pub country: String,
-        pub state: Option<String>,
-        pub zip: String,
-    }
-    impl Address {
-        pub fn decrypt(&self, key: &str, iv: &str) -> anyhow::Result<Address> {
-            Ok(Address {
-                id: self.id,
-                street: decrypt((key, iv), &self.street)?,
-                city: decrypt((key, iv), &self.city)?,
-                country: decrypt((key, iv), &self.country)?,
-                state: if let Some(state) = &self.state {
-                    Some(decrypt((key, iv), state)?)
-                } else {
-                    None
-                },
-                zip: decrypt((key, iv), &self.zip)?,
-            })
-        }
-    }
-
-    #[derive(cynic::InputObject, Debug, Clone)]
-    pub struct PaymentCardIn {
-        pub iv: String,
-        pub name: String,
-        pub name_on_card: String,
-        pub number: String,
-        pub cvv: String,
-        pub expiry: ExpiryIn,
-        pub color: Option<String>,
-        pub billing_address: Option<AddressIn>,
-    }
-
-    impl PaymentCardIn {
-        pub fn encrypt(&self, key: &str) -> PaymentCardIn {
-            PaymentCardIn {
-                iv: self.iv.clone(),
-                name: self.name.clone(),
-                expiry: self.expiry.clone(),
-                number: encrypt(key, &self.iv, &self.number),
-                name_on_card: encrypt(key, &self.iv, &self.name_on_card),
-                cvv: encrypt(key, &self.iv, &self.cvv),
-                color: if let Some(color) = &self.color {
-                    Some(encrypt(key, &self.iv, color))
-                } else {
-                    None
-                },
-                billing_address: if let Some(address) = &self.billing_address {
-                    Some(address.encrypt(key, &self.iv))
-                } else {
-                    None
-                },
-            }
-        }
-    }
-
-    #[derive(cynic::QueryFragment, Debug, Clone)]
-    pub struct PaymentCard {
-        pub id: i32,
-        pub iv: String,
-        pub name: String,
-        pub name_on_card: String,
-        pub number: String,
-        pub cvv: String,
-        pub expiry: Expiry,
-        pub color: Option<String>,
-        pub billing_address: Option<Address>,
-    }
-    impl PaymentCard {
-        pub fn last_four(&self) -> String {
-            self.number
-                .chars()
-                .skip(self.number.len() - 4)
-                .take(4)
-                .collect::<String>()
-        }
-        pub fn color_str(&self) -> String {
-            if let Some(color) = &self.color {
-                color.clone()
-            } else {
-                "".to_string()
-            }
-        }
-        pub fn expiry_str(&self) -> String {
-            self.expiry.to_string()
-        }
-        pub fn decrypt(&self, key: &str) -> anyhow::Result<PaymentCard> {
-            Ok(PaymentCard {
-                id: self.id,
-                iv: self.iv.clone(),
-                name: self.name.clone(),
-                expiry: self.expiry.clone(),
-                number: decrypt((key, &self.iv), &self.number)?,
-                name_on_card: decrypt((key, &self.iv), &self.name_on_card)?,
-                cvv: decrypt((key, &self.iv), &self.cvv)?,
-                color: if let Some(color) = &self.color {
-                    Some(decrypt((key, &self.iv), color)?)
-                } else {
-                    None
-                },
-                billing_address: if let Some(address) = &self.billing_address {
-                    Some(address.decrypt(key, &self.iv)?)
-                } else {
-                    None
-                },
-            })
-        }
-    }
-    #[derive(cynic::InputObject, Debug, Clone)]
-    pub struct AddPaymentCardIn {
-        pub payment: PaymentCardIn,
-        pub vault_id: Option<i32>,
-    }
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct AddPaymentCardMutationVariables {
-        pub input: AddPaymentCardIn,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "AddPaymentCardMutationVariables"
-    )]
-    pub struct AddPaymentCardMutation {
-        #[arguments(input = AddPaymentCardIn {
-            payment: args.input.payment.clone(),
-            vault_id: args.input.vault_id
-        })]
-        pub add_payment_card: PaymentCard,
-    }
-
-    #[derive(cynic::QueryFragment, Debug, Clone)]
-    pub struct Note {
-        pub id: i32,
-        pub iv: String,
-        pub title: String,
-        pub content: String,
-        pub created: Date,
-        pub modified: Option<Date>,
-    }
-    impl Note {
-        pub fn decrypt(&self, key: &str) -> anyhow::Result<Note> {
-            Ok(Note {
-                id: self.id,
-                iv: self.iv.clone(),
-                title: decrypt((key, &self.iv), &self.title)?,
-                content: decrypt((key, &self.iv), &self.content)?,
-                created: self.created.clone(),
-                modified: if let Some(modified) = &self.modified {
-                    Some(modified.clone())
-                } else {
-                    None
-                },
-            })
-        }
-    }
-
-    #[derive(cynic::FragmentArguments, cynic::InputObject, Debug, Clone)]
-    pub struct NoteIn {
-        pub iv: String,
-        pub title: String,
-        pub content: String,
-        pub vault_id: Option<i32>,
-    }
-    impl NoteIn {
-        pub fn encrypt(&self, key: &str) -> NoteIn {
-            NoteIn {
-                iv: self.iv.clone(),
-                title: encrypt(key, &self.iv, &self.title),
-                content: encrypt(key, &self.iv, &self.content),
-                vault_id: self.vault_id,
-            }
-        }
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(graphql_type = "Mutation", argument_struct = "NoteIn")]
-    pub struct AddNoteMutation {
-        #[arguments(input = NoteIn {
-            iv: args.iv.clone(),
-            title: args.title.clone(),
-            content: args.content.clone(),
-            vault_id: args.vault_id
-        })]
-        pub add_note: Note,
-    }
-
-    #[derive(cynic::FragmentArguments, Debug)]
-    pub struct DeleteNoteMutationVariables {
-        pub id: i32,
-    }
-
-    #[derive(cynic::QueryFragment, Debug)]
-    #[cynic(
-        graphql_type = "Mutation",
-        argument_struct = "DeleteNoteMutationVariables"
-    )]
-    pub struct DeleteNoteMutation {
-        #[arguments(id = args.id)]
-        pub delete_note: i32,
-    }
-}
-
-mod schema {
-    cynic::use_schema!(r#"src/schema.graphql"#);
-}
 fn new_request(access_token: &str) -> reqwest::RequestBuilder {
     reqwest::Client::new()
         .post(API_ENDPOINT)
@@ -501,7 +17,7 @@ fn new_request(access_token: &str) -> reqwest::RequestBuilder {
 pub async fn run_me_query(
     access_token: &str,
     grep: Option<String>,
-) -> cynic::GraphQlResponse<queries::MeQuery> {
+) -> cynic::GraphQlResponse<MeQuery> {
     let operation = build_me_query(grep);
 
     new_request(access_token)
@@ -510,14 +26,14 @@ pub async fn run_me_query(
         .unwrap()
 }
 
-fn build_me_query(grep: Option<String>) -> cynic::Operation<'static, queries::MeQuery> {
+fn build_me_query(grep: Option<String>) -> cynic::Operation<'static, MeQuery> {
     use cynic::QueryBuilder;
-    queries::MeQuery::build(queries::CredentialsQueryVariables { grep })
+    MeQuery::build(CredentialsQueryVariables { grep })
 }
 
 pub async fn run_payment_card_query(
     access_token: &str,
-) -> cynic::GraphQlResponse<queries::PaymentCardMeQuery> {
+) -> cynic::GraphQlResponse<PaymentCardMeQuery> {
     let operation = build_payment_card_query();
     new_request(access_token)
         .run_graphql(operation)
@@ -525,12 +41,12 @@ pub async fn run_payment_card_query(
         .unwrap()
 }
 
-fn build_payment_card_query() -> cynic::Operation<'static, queries::PaymentCardMeQuery> {
+fn build_payment_card_query() -> cynic::Operation<'static, PaymentCardMeQuery> {
     use cynic::QueryBuilder;
-    queries::PaymentCardMeQuery::build(queries::EmptyQueryVariables {})
+    PaymentCardMeQuery::build(EmptyQueryVariables {})
 }
 
-pub async fn run_notes_query(access_token: &str) -> cynic::GraphQlResponse<queries::NotesMeQuery> {
+pub async fn run_notes_query(access_token: &str) -> cynic::GraphQlResponse<NotesMeQuery> {
     let operation = build_notes_query();
     new_request(access_token)
         .run_graphql(operation)
@@ -538,14 +54,12 @@ pub async fn run_notes_query(access_token: &str) -> cynic::GraphQlResponse<queri
         .unwrap()
 }
 
-fn build_notes_query() -> cynic::Operation<'static, queries::NotesMeQuery> {
+fn build_notes_query() -> cynic::Operation<'static, NotesMeQuery> {
     use cynic::QueryBuilder;
-    queries::NotesMeQuery::build(queries::EmptyQueryVariables {})
+    NotesMeQuery::build(EmptyQueryVariables {})
 }
 
-pub async fn run_plain_me_query(
-    access_token: &str,
-) -> cynic::GraphQlResponse<queries::PlainMeQuery> {
+pub async fn run_plain_me_query(access_token: &str) -> cynic::GraphQlResponse<PlainMeQuery> {
     let operation = build_plain_me_query();
     new_request(access_token)
         .run_graphql(operation)
@@ -553,16 +67,16 @@ pub async fn run_plain_me_query(
         .unwrap()
 }
 
-fn build_plain_me_query() -> cynic::Operation<'static, queries::PlainMeQuery> {
+fn build_plain_me_query() -> cynic::Operation<'static, PlainMeQuery> {
     use cynic::QueryBuilder;
-    queries::PlainMeQuery::build(queries::EmptyQueryVariables {})
+    PlainMeQuery::build(EmptyQueryVariables {})
 }
 
 pub async fn run_add_credentials_group_mutation(
     access_token: &str,
-    credentials: Vec<queries::CredentialsIn>,
+    credentials: Vec<CredentialsIn>,
     vault_id: Option<i32>,
-) -> cynic::GraphQlResponse<queries::AddGredentialsGroupMutation> {
+) -> cynic::GraphQlResponse<AddGredentialsGroupMutation> {
     let operation = build_add_credentials_group_mutation(credentials, vault_id);
 
     new_request(access_token)
@@ -572,14 +86,13 @@ pub async fn run_add_credentials_group_mutation(
 }
 
 fn build_add_credentials_group_mutation(
-    credentials: Vec<queries::CredentialsIn>,
+    credentials: Vec<CredentialsIn>,
     vault_id: Option<i32>,
-) -> cynic::Operation<'static, queries::AddGredentialsGroupMutation> {
+) -> cynic::Operation<'static, AddGredentialsGroupMutation> {
     use cynic::MutationBuilder;
-    use queries::{AddGredentialsGroupMutation, AddGredentialsGroupMutationVariables};
 
     AddGredentialsGroupMutation::build(&AddGredentialsGroupMutationVariables {
-        input: queries::AddCredentialsGroupIn {
+        input: AddCredentialsGroupIn {
             credentials,
             vault_id,
         },
@@ -590,7 +103,7 @@ pub async fn run_delete_credentials_mutation(
     access_token: &str,
     grep: &str,
     index: Option<i32>,
-) -> cynic::GraphQlResponse<queries::DeleteCredentialsMutation> {
+) -> cynic::GraphQlResponse<DeleteCredentialsMutation> {
     let operation = build_delete_credentials_mutation(grep.into(), index);
     new_request(access_token)
         .run_graphql(operation)
@@ -601,12 +114,11 @@ pub async fn run_delete_credentials_mutation(
 fn build_delete_credentials_mutation(
     grep: String,
     index: Option<i32>,
-) -> cynic::Operation<'static, queries::DeleteCredentialsMutation> {
+) -> cynic::Operation<'static, DeleteCredentialsMutation> {
     use cynic::MutationBuilder;
-    use queries::{DeleteCredentialsMutation, DeleteCredentialsMutationVariables};
 
     DeleteCredentialsMutation::build(&DeleteCredentialsMutationVariables {
-        input: queries::DeleteCredentialsIn { grep, index },
+        input: DeleteCredentialsIn { grep, index },
     })
 }
 
@@ -614,7 +126,7 @@ pub async fn run_migrate_mutation(
     access_token: &str,
     old_key: &str,
     new_key: &str,
-) -> cynic::GraphQlResponse<queries::MigrateMutation> {
+) -> cynic::GraphQlResponse<MigrateMutation> {
     let operation = build_migrate_mutation(old_key, new_key);
     new_request(access_token)
         .run_graphql(operation)
@@ -625,9 +137,8 @@ pub async fn run_migrate_mutation(
 fn build_migrate_mutation(
     old_key: &str,
     new_key: &str,
-) -> cynic::Operation<'static, queries::MigrateMutation> {
+) -> cynic::Operation<'static, MigrateMutation> {
     use cynic::MutationBuilder;
-    use queries::{MigrateMutation, MigrateMutationVariables};
 
     MigrateMutation::build(&MigrateMutationVariables {
         old_key: String::from(old_key),
@@ -637,10 +148,10 @@ fn build_migrate_mutation(
 
 pub async fn run_add_payment_card_mutation(
     access_token: &str,
-    payment: queries::PaymentCardIn,
+    payment: PaymentCardIn,
     vault_id: Option<i32>,
-) -> cynic::GraphQlResponse<queries::AddPaymentCardMutation> {
-    let operation: cynic::Operation<queries::AddPaymentCardMutation> =
+) -> cynic::GraphQlResponse<AddPaymentCardMutation> {
+    let operation: cynic::Operation<AddPaymentCardMutation> =
         build_add_payment_card_mutation(payment, vault_id);
 
     new_request(access_token)
@@ -650,22 +161,21 @@ pub async fn run_add_payment_card_mutation(
 }
 
 fn build_add_payment_card_mutation(
-    payment: queries::PaymentCardIn,
+    payment: PaymentCardIn,
     vault_id: Option<i32>,
-) -> cynic::Operation<'static, queries::AddPaymentCardMutation> {
+) -> cynic::Operation<'static, AddPaymentCardMutation> {
     use cynic::MutationBuilder;
-    use queries::{AddPaymentCardMutation, AddPaymentCardMutationVariables};
 
     AddPaymentCardMutation::build(&AddPaymentCardMutationVariables {
-        input: queries::AddPaymentCardIn { payment, vault_id },
+        input: AddPaymentCardIn { payment, vault_id },
     })
 }
 
 pub async fn run_delete_payment_card_mutation(
     access_token: &str,
     id: i32,
-) -> cynic::GraphQlResponse<queries::DeletePaymentCardMutation> {
-    let operation: cynic::Operation<queries::DeletePaymentCardMutation> =
+) -> cynic::GraphQlResponse<DeletePaymentCardMutation> {
+    let operation: cynic::Operation<DeletePaymentCardMutation> =
         build_delete_payment_card_mutation(id);
 
     new_request(access_token)
@@ -676,18 +186,17 @@ pub async fn run_delete_payment_card_mutation(
 
 fn build_delete_payment_card_mutation(
     id: i32,
-) -> cynic::Operation<'static, queries::DeletePaymentCardMutation> {
+) -> cynic::Operation<'static, DeletePaymentCardMutation> {
     use cynic::MutationBuilder;
-    use queries::{DeletePaymentCardMutation, DeletePaymentCardMutationVariables};
 
     DeletePaymentCardMutation::build(&DeletePaymentCardMutationVariables { id })
 }
 
 pub async fn run_add_note_mutation(
     access_token: &str,
-    note: &queries::NoteIn,
-) -> cynic::GraphQlResponse<queries::AddNoteMutation> {
-    let operation: cynic::Operation<queries::AddNoteMutation> = build_add_note_mutation(note);
+    note: &NoteIn,
+) -> cynic::GraphQlResponse<AddNoteMutation> {
+    let operation: cynic::Operation<AddNoteMutation> = build_add_note_mutation(note);
 
     new_request(access_token)
         .run_graphql(operation)
@@ -695,11 +204,8 @@ pub async fn run_add_note_mutation(
         .unwrap()
 }
 
-fn build_add_note_mutation(
-    note: &queries::NoteIn,
-) -> cynic::Operation<'static, queries::AddNoteMutation> {
+fn build_add_note_mutation(note: &NoteIn) -> cynic::Operation<'static, AddNoteMutation> {
     use cynic::MutationBuilder;
-    use queries::AddNoteMutation;
 
     AddNoteMutation::build(note)
 }
@@ -707,7 +213,7 @@ fn build_add_note_mutation(
 pub async fn run_delete_note_mutation(
     access_token: &str,
     id: i32,
-) -> cynic::GraphQlResponse<queries::DeleteNoteMutation> {
+) -> cynic::GraphQlResponse<DeleteNoteMutation> {
     let operation = build_delete_note_mutation(id);
     new_request(access_token)
         .run_graphql(operation)
@@ -715,9 +221,8 @@ pub async fn run_delete_note_mutation(
         .unwrap()
 }
 
-fn build_delete_note_mutation(id: i32) -> cynic::Operation<'static, queries::DeleteNoteMutation> {
+fn build_delete_note_mutation(id: i32) -> cynic::Operation<'static, DeleteNoteMutation> {
     use cynic::MutationBuilder;
-    use queries::{DeleteNoteMutation, DeleteNoteMutationVariables};
 
     DeleteNoteMutation::build(&DeleteNoteMutationVariables { id })
 }
